removed addcontacts controller, replaced with alertview
