+++ ghost/ghost/Base.lproj/Main.storyboard
@@ -32,7 +32,7 @@
                     <navigationItem key="navigationItem" id="ULf-Vs-Arw">
                         <barButtonItem key="rightBarButtonItem" systemItem="add" id="MHb-dM-t7h">
                             <connections>
-                                <segue destination="ut8-YL-1Ci" kind="show" identifier="to-add-contact" id="7Md-rP-eKW"/>
+                                <action selector="addContact:" destination="MMZ-d9-1dU" id="MKI-f1-AE9"/>
                             </connections>
                         </barButtonItem>
                     </navigationItem>
@@ -40,63 +40,6 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="tAD-2U-XCd" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="2643" y="49"/>
-        </scene>
-        <!--Add Contact View Controller-->
-        <scene sceneID="Wh4-yc-NHw">
-            <objects>
-                <viewController id="ut8-YL-1Ci" customClass="AddContactViewController" customModule="ghost" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="nci-Z1-Wwq"/>
-                        <viewControllerLayoutGuide type="bottom" id="MNT-U6-fae"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="pFq-Zj-cdK">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="Oyd-hB-6fl">
-                                <rect key="frame" x="221" y="181" width="159" height="30"/>
-                                <constraints>
-                                    <constraint firstAttribute="width" constant="159" id="WAw-Ft-IQi"/>
-                                    <constraint firstAttribute="height" constant="30" id="jbv-0G-DMY"/>
-                                </constraints>
-                                <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                                <textInputTraits key="textInputTraits"/>
-                            </textField>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="SpJ-iT-yUs">
-                                <rect key="frame" x="276" y="238" width="49" height="30"/>
-                                <state key="normal" title="Submit"/>
-                                <connections>
-                                    <action selector="submitContact:" destination="ut8-YL-1Ci" eventType="touchUpInside" id="5IK-Mw-0gj"/>
-                                </connections>
-                            </button>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Enter Username" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="K1K-xc-i3A">
-                                <rect key="frame" x="238" y="124" width="124" height="21"/>
-                                <constraints>
-                                    <constraint firstAttribute="width" constant="124" id="Cdd-je-W6K"/>
-                                    <constraint firstAttribute="height" constant="21" id="fEO-1h-0pE"/>
-                                </constraints>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <color key="textColor" red="1" green="0.49803921569999998" blue="0.0039215686269999999" alpha="1" colorSpace="calibratedRGB"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <constraints>
-                            <constraint firstItem="K1K-xc-i3A" firstAttribute="top" secondItem="nci-Z1-Wwq" secondAttribute="bottom" constant="60" id="8mf-8B-g0m"/>
-                            <constraint firstItem="Oyd-hB-6fl" firstAttribute="top" secondItem="K1K-xc-i3A" secondAttribute="bottom" constant="36" id="Jo4-Ox-WMX"/>
-                            <constraint firstItem="K1K-xc-i3A" firstAttribute="centerX" secondItem="pFq-Zj-cdK" secondAttribute="centerX" id="OeA-Gg-TXe"/>
-                            <constraint firstItem="SpJ-iT-yUs" firstAttribute="top" secondItem="Oyd-hB-6fl" secondAttribute="bottom" constant="27" id="VIR-Ud-wLO"/>
-                            <constraint firstItem="K1K-xc-i3A" firstAttribute="centerX" secondItem="SpJ-iT-yUs" secondAttribute="centerX" id="u09-Su-fdc"/>
-                            <constraint firstItem="K1K-xc-i3A" firstAttribute="centerX" secondItem="Oyd-hB-6fl" secondAttribute="centerX" id="uUe-8F-G8p"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="usernameText" destination="Oyd-hB-6fl" id="NJJ-lv-1GY"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="uno-8D-c8F" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="3399" y="49"/>
         </scene>
         <!--Main Table View Controller-->
         <scene sceneID="pfa-zd-d1x">
